fe add authorization

+++ server/src/resolvers/Mutation.ts
@@ -6,15 +6,42 @@ import { MutationResolvers } from "../generated/graphqlgen";
 import * as bcrypt from 'bcryptjs';
 import { APP_SECRET } from '../utils';
 import { sign } from 'jsonwebtoken';
+import { getUserId } from "../utils";
 
 export const Mutation: MutationResolvers.Type = {
   ...MutationResolvers.defaultResolvers,
-  createPost: (parent, args, ctx) => {
-    return ctx.prisma.createPost(args.input);
+  createDraft: (parent, args, ctx) => {
+    const userId = getUserId(ctx);
+    return ctx.prisma.createPost({
+      ...args.input,
+      author: { connect: { id: userId } },
+    });
   },
-  deletePost: (parent, args, ctx) => {
+  deletePost: async (parent, args, ctx) => {
+    const userId = getUserId(ctx);
+    const postExists = await ctx.prisma.$exists.post({
+      ...args,
+      author: { id: userId }
+    });
+    if (!postExists) {
+      throw new Error('Post not found or you are not author!');
+    }
     return ctx.prisma.deletePost({
       id: args.id
+    });
+  },
+  publish: async (parent, args, ctx) => {
+    const userId = getUserId(ctx);
+    const postExists = await ctx.prisma.$exists.post({
+      ...args,
+      author: { id: userId }
+    });
+    if (!postExists) {
+      throw new Error('Post not found or you are not author!');
+    }
+    return ctx.prisma.updatePost({
+      where: { id: args.id },
+      data: { published: true }
     });
   },
   signup: async (parent, args, ctx) => {
